单机h2版

+++ src/main/java/com/chester/svc/auth/web/model/req/ReqRule.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 @Data
 public class ReqRule {
-    Integer ruleId;
+    Long ruleId;
     List<String> roles;
 }
